+++ packages/backend/src/models/ProjectModel/ProjectModel.mock.ts
@@ -125,6 +125,9 @@ export const lightdashConfigMock: LightdashConfig = {
         cacheStateTimeSeconds: 86400,
         s3: {},
     },
+    groups: {
+        enabled: false,
+    },
 };
 
 const dbtCloudIDEProjectConfigMock: DbtCloudIDEProjectConfig = {