+++ packages/cli/src/handlers/generate.ts
@@ -96,6 +96,7 @@ export const generateHandler = async (options: GenerateHandlerOptions) => {
         target: options.target,
         select: options.select || options.models,
         exclude: options.exclude,
+        vars: options.vars || undefined,
     });
 
     GlobalState.debug(`> Compiled models: ${compiledModels.length}`);
@@ -117,6 +118,7 @@ export const generateHandler = async (options: GenerateHandlerOptions) => {
                     docs: manifest.docs,
                     includeMeta: !options.excludeMeta,
                     projectDir: absoluteProjectPath,
+                    projectName: context.projectName,
                 },
             );
             try {