+++ packages/frontend/src/components/Explorer/CustomMetricModal/index.tsx
@@ -1,11 +1,17 @@
 import {
+    canApplyFormattingToCustomMetric,
+    CustomMetricFormat,
+    Dimension,
     fieldId as getFieldId,
     friendlyName,
     isAdditionalMetric,
     isDimension,
     MetricType,
+    NumberSeparator,
+    // NOTE: The types are the same, but we can't derive enums from other enums
+    TableCalculationFormatType as CustomMetricFormatType,
 } from '@lightdash/common';
-import { isNumericItem } from '@lightdash/common/src/utils/item';
+
 import {
     Accordion,
     Button,
@@ -19,7 +25,9 @@ import {
 import { useForm } from '@mantine/form';
 import { useFeatureFlagEnabled } from 'posthog-js/react';
 import { useEffect, useMemo, useState } from 'react';
+import { ValueOf } from 'type-fest';
 import { v4 as uuidv4 } from 'uuid';
+import { FormatForm } from '../../../features/tableCalculation/components/FormatForm';
 import useToaster from '../../../hooks/toaster/useToaster';
 import { useExplore } from '../../../hooks/useExplore';
 import { useExplorerContext } from '../../../providers/ExplorerProvider';
@@ -61,6 +69,22 @@ export const CustomMetricModal = () => {
 
     const { data: exploreData } = useExplore(tableName);
 
+    let dimensionToCheck: Dimension | undefined;
+
+    if (isDimension(item)) {
+        dimensionToCheck = item;
+    }
+    if (isEditing && isAdditionalMetric(item) && item.baseDimensionName) {
+        dimensionToCheck =
+            exploreData?.tables[item.table]?.dimensions[item.baseDimensionName];
+    }
+
+    const canApplyFormatting =
+        isCustomMetricFormattingEnabled &&
+        dimensionToCheck &&
+        customMetricType &&
+        canApplyFormattingToCustomMetric(dimensionToCheck, customMetricType);
+
     const additionalMetrics = useExplorerContext(
         (context) =>
             context.state.unsavedChartVersion.metricQuery.additionalMetrics,
@@ -73,6 +97,15 @@ export const CustomMetricModal = () => {
         initialValues: {
             customMetricLabel: '',
             percentile: 50,
+            format: {
+                type: CustomMetricFormatType.DEFAULT,
+                round: undefined,
+                separator: NumberSeparator.DEFAULT,
+                currency: 'USD',
+                compact: undefined,
+                prefix: undefined,
+                suffix: undefined,
+            },
         },
         validate: {
             customMetricLabel: (label) => {
@@ -205,6 +238,14 @@ export const CustomMetricModal = () => {
         }
     }, [isEditing, item]);
 
+    const getFormatInputProps = (path: keyof CustomMetricFormat) =>
+        form.getInputProps(`format.${path}`);
+
+    const setFormatFieldValue = (
+        path: keyof CustomMetricFormat,
+        value: ValueOf<CustomMetricFormat>,
+    ) => form.setFieldValue(`format.${path}`, value);
+
     return item ? (
         <Modal
             size="xl"
@@ -236,16 +277,24 @@ export const CustomMetricModal = () => {
                         />
                     )}
                     <Accordion chevronPosition="left" chevronSize="xs">
-                        {isCustomMetricFormattingEnabled &&
-                            isNumericItem(item) && (
-                                <Accordion.Item value="format">
-                                    <Accordion.Control>
-                                        <Text fw={500} fz="sm">
-                                            Format
-                                        </Text>
-                                    </Accordion.Control>
-                                </Accordion.Item>
-                            )}
+                        {canApplyFormatting && (
+                            <Accordion.Item value="format">
+                                <Accordion.Control>
+                                    <Text fw={500} fz="sm">
+                                        Format
+                                    </Text>
+                                </Accordion.Control>
+                                <Accordion.Panel>
+                                    <FormatForm
+                                        formatInputProps={getFormatInputProps}
+                                        format={form.values.format}
+                                        setFormatFieldValue={
+                                            setFormatFieldValue
+                                        }
+                                    />
+                                </Accordion.Panel>
+                            </Accordion.Item>
+                        )}
                         <Accordion.Item value="filters">
                             <Accordion.Control>
                                 <Text fw={500} fz="sm">