+++ packages/frontend/src/components/common/Filters/FilterInputs/DateFilterInputs.tsx
@@ -13,6 +13,7 @@ import {
     TimeFrames,
 } from '@lightdash/common';
 import { Flex, NumberInput, Text } from '@mantine/core';
+import moment from 'moment';
 import { FilterInputsProps } from '.';
 import { useFiltersContext } from '../FiltersProvider';
 import { getFirstDayOfWeek } from '../utils/filterDateUtils';
@@ -176,6 +177,12 @@ const DateFilterInputs = <T extends ConditionalRule = DateFilterRule>(
                         // @ts-ignore
                         placeholder={placeholder}
                         withSeconds
+                        valueFormat={
+                            rule.values &&
+                            moment(rule.values[0])
+                                .utc()
+                                .format('DD/MM/YYYY HH:mm:ss')
+                        }
                         // FIXME: mantine v7
                         // mantine does not set the first day of the week based on the locale
                         // so we need to do it manually and always pass it as a prop
@@ -201,7 +208,7 @@ const DateFilterInputs = <T extends ConditionalRule = DateFilterRule>(
                                         : [
                                               formatTimestamp(
                                                   value,
-                                                  TimeFrames.MILLISECOND,
+                                                  TimeFrames.SECOND,
                                               ),
                                           ],
                             });