+++ packages/frontend/src/components/common/ResourceView/ResourceViewGrid/index.tsx
@@ -37,6 +37,7 @@ type ResourceViewGridProps = ResourceViewGridCommonProps &
 
 type DraggableItemProps = Pick<ResourceViewGridProps, 'onAction'> & {
     item: ResourceViewItem;
+    allowDelete?: boolean;
     index: number;
     onAction: (newAction: ResourceViewItemActionState) => void;
     projectUuid: string;
@@ -45,6 +46,7 @@ type DraggableItemProps = Pick<ResourceViewGridProps, 'onAction'> & {
 
 const DraggableItem: FC<DraggableItemProps> = ({
     item,
+    allowDelete,
     index,
     onAction,
     projectUuid,
@@ -94,18 +96,21 @@ const DraggableItem: FC<DraggableItemProps> = ({
                         {item.type === ResourceViewItemType.SPACE ? (
                             <ResourceViewGridSpaceItem
                                 item={item}
+                                allowDelete={allowDelete}
                                 onAction={onAction}
                                 dragIcon={DragIcon}
                             />
                         ) : item.type === ResourceViewItemType.DASHBOARD ? (
                             <ResourceViewGridDashboardItem
                                 item={item}
+                                allowDelete={allowDelete}
                                 onAction={onAction}
                                 dragIcon={DragIcon}
                             />
                         ) : item.type === ResourceViewItemType.CHART ? (
                             <ResourceViewGridChartItem
                                 item={item}
+                                allowDelete={allowDelete}
                                 onAction={onAction}
                                 dragIcon={DragIcon}
                             />
@@ -134,7 +139,7 @@ const ResourceViewGrid: FC<ResourceViewGridProps> = ({
     onAction,
     hasReorder = false,
 }) => {
-    const { reorderItems } = usePinnedItemsContext();
+    const { reorderItems, allowDelete } = usePinnedItemsContext();
     const { projectUuid } = useParams<{ projectUuid: string }>();
 
     const groupedItems = useMemo(() => {
@@ -226,6 +231,7 @@ const ResourceViewGrid: FC<ResourceViewGridProps> = ({
                                                 item.type + '-' + item.data.uuid
                                             }
                                             item={item}
+                                            allowDelete={allowDelete}
                                             index={index}
                                             onAction={onAction}
                                             projectUuid={projectUuid}